# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Reliastra MVP API",
    "description": "External dependency intelligence platform API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/verify/{verification_id}": {
      "get": {
        "tags": [
          "Evidence Verification"
        ],
        "summary": "Verify Evidence",
        "operationId": "verify_evidence_v1_verify__verification_id__get",
        "parameters": [
          {
            "name": "verification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Verification Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/verify/{verification_id}/hash": {
      "get": {
        "tags": [
          "Evidence Verification"
        ],
        "summary": "Verify Evidence Hash",
        "operationId": "verify_evidence_hash_v1_verify__verification_id__hash_get",
        "parameters": [
          {
            "name": "verification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Verification Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationHashResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/verify/{verification_id}/evidence": {
      "get": {
        "tags": [
          "Evidence Verification"
        ],
        "summary": "Get Verification Evidence",
        "operationId": "get_verification_evidence_v1_verify__verification_id__evidence_get",
        "parameters": [
          {
            "name": "verification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Verification Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationEvidenceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/public/vendors/{vendor_name}": {
      "get": {
        "tags": [
          "Public Vendors"
        ],
        "summary": "Get Public Vendor",
        "operationId": "get_public_vendor_v1_public_vendors__vendor_name__get",
        "parameters": [
          {
            "name": "vendor_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Vendor Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/public/vendors/{vendor_name}/history": {
      "get": {
        "tags": [
          "Public Vendors"
        ],
        "summary": "Get Public Vendor History",
        "operationId": "get_public_vendor_history_v1_public_vendors__vendor_name__history_get",
        "parameters": [
          {
            "name": "vendor_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Vendor Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/public/vendors/{slug}/incidents": {
      "get": {
        "tags": [
          "Public Vendors"
        ],
        "summary": "Get Public Vendor Incidents",
        "operationId": "get_public_vendor_incidents_v1_public_vendors__slug__incidents_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VendorIncidentResponse"
                  },
                  "title": "Response Get Public Vendor Incidents V1 Public Vendors  Slug  Incidents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VendorDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "vendor_name": {
            "type": "string",
            "title": "Vendor Name"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "is_public": {
            "type": "boolean",
            "title": "Is Public"
          },
          "last_check_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Check At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "recent_status": {
            "type": "string",
            "title": "Recent Status",
            "default": "operational"
          }
        },
        "type": "object",
        "required": [
          "id",
          "vendor_name",
          "display_name",
          "category",
          "is_public",
          "created_at",
          "updated_at"
        ],
        "title": "VendorDetailResponse"
      },
      "VendorHistoryResponse": {
        "properties": {
          "vendor_name": {
            "type": "string",
            "title": "Vendor Name"
          },
          "uptime_percentage_24h": {
            "type": "number",
            "title": "Uptime Percentage 24H",
            "default": 99.99
          },
          "avg_latency_ms_24h": {
            "type": "number",
            "title": "Avg Latency Ms 24H",
            "default": 45.0
          },
          "recent_checks_count": {
            "type": "integer",
            "title": "Recent Checks Count",
            "default": 288
          }
        },
        "type": "object",
        "required": [
          "vendor_name"
        ],
        "title": "VendorHistoryResponse"
      },
      "VendorIncidentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "vendor_id": {
            "type": "string",
            "format": "uuid",
            "title": "Vendor Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "vendor_id",
          "title",
          "description",
          "status",
          "started_at",
          "resolved_at",
          "created_at",
          "updated_at"
        ],
        "title": "VendorIncidentResponse"
      },
...
      "VerificationEvidenceResponse": {
        "properties": {
          "verification_id": {
            "type": "string",
            "title": "Verification Id"
          },
          "incident_id": {
            "type": "string",
            "format": "uuid",
            "title": "Incident Id"
          },
          "dependency_id": {
            "type": "string",
            "format": "uuid",
            "title": "Dependency Id"
          },
          "time_window_start": {
            "type": "string",
            "format": "date-time",
            "title": "Time Window Start"
          },
          "time_window_end": {
            "type": "string",
            "format": "date-time",
            "title": "Time Window End"
          },
          "methodology_version": {
            "type": "string",
            "title": "Methodology Version"
          },
          "data_hash": {
            "type": "string",
            "title": "Data Hash"
          },
          "observations": {
            "items": {},
            "type": "array",
            "title": "Observations"
          },
          "attribution_result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Attribution Result"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "verification_id",
          "incident_id",
          "dependency_id",
          "time_window_start",
          "time_window_end",
          "methodology_version",
          "data_hash",
          "observations",
          "attribution_result",
          "created_at"
        ],
        "title": "VerificationEvidenceResponse"
      },
      "VerificationHashResponse": {
        "properties": {
          "verification_id": {
            "type": "string",
            "title": "Verification Id"
          },
          "data_hash": {
            "type": "string",
            "title": "Data Hash"
          },
          "report_checksum": {
            "type": "string",
            "title": "Report Checksum"
          }
        },
        "type": "object",
        "required": [
          "verification_id",
          "data_hash",
          "report_checksum"
        ],
        "title": "VerificationHashResponse"
      },
      "VerificationResponse": {
        "properties": {
          "verification_id": {
            "type": "string",
            "title": "Verification Id"
          },
          "incident_id": {
            "type": "string",
            "format": "uuid",
            "title": "Incident Id"
          },
          "dependency_id": {
            "type": "string",
            "format": "uuid",
            "title": "Dependency Id"
          },
          "org_id": {
            "type": "string",
            "format": "uuid",
            "title": "Org Id"
          },
          "time_window_start": {
            "type": "string",
            "format": "date-time",
            "title": "Time Window Start"
          },
          "time_window_end": {
            "type": "string",
            "format": "date-time",
            "title": "Time Window End"
          },
          "methodology_version": {
            "type": "string",
            "title": "Methodology Version"
          },
          "data_hash": {
            "type": "string",
            "title": "Data Hash"
          },
          "report_checksum": {
            "type": "string",
            "title": "Report Checksum"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "hashes_match": {
            "type": "boolean",
            "title": "Hashes Match"
          }
        },
        "type": "object",
        "required": [
          "verification_id",
          "incident_id",
          "dependency_id",
          "org_id",
          "time_window_start",
          "time_window_end",
          "methodology_version",
          "data_hash",
          "report_checksum",
          "created_at",
          "hashes_match"
        ],
        "title": "VerificationResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}